split lextwt.go into smaller portions.
